# docs/customizations/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Mayan EDMS API",
    "description": "Free Open Source Electronic Document Management System",
    "license": {
      "name": "Apache 2.0"
    },
    "version": "v4"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "basePath": "/api/v4",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Basic": {
      "type": "basic"
    }
  },
  "security": [
    {
      "Basic": []
    }
  ],
  "paths": {
...
    "/cabinets/": {
      "get": {
        "operationId": "cabinets_list",
        "description": "Returns a list of all the cabinets.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cabinets"
        ]
      },
      "post": {
        "operationId": "cabinets_create",
        "description": "Create a new cabinet.",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cabinets"
        ]
      },
      "parameters": []
    },
    "/cabinets/{cabinet_id}/": {
      "get": {
        "operationId": "cabinets_read",
        "description": "Returns the details of the selected cabinet.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cabinets"
        ]
      },
      "put": {
        "operationId": "cabinets_update",
        "description": "Edit the selected cabinet.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cabinets"
        ]
      },
      "patch": {
        "operationId": "cabinets_partial_update",
        "description": "Edit the selected cabinet.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cabinets"
        ]
      },
      "delete": {
        "operationId": "cabinets_delete",
        "description": "Delete the selected cabinet.",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "cabinets"
        ]
      },
      "parameters": [
        {
          "name": "cabinet_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
    "/cabinets/{cabinet_id}/documents/": {
      "get": {
        "operationId": "cabinets_documents_list",
        "description": "Returns a list of all the documents contained in a particular cabinet.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cabinets"
        ]
      },
      "parameters": [
        {
          "name": "cabinet_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/documents/{document_id}/cabinets/": {
      "get": {
        "operationId": "documents_cabinets_list",
        "description": "Returns a list of all the cabinets to which a document belongs.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "parameters": [
        {
          "name": "document_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
  },
...
}